Retrofitted static_class_info() in tests.
The tests used a simple hash that wasn't inherited from DBIx::NinjaORM::static_class_info()
in their local static_class_info() declarations. The use of a hash isn't a problem per se,
but the lack of inheritance means the introduction of objects upstream didn't result in
the tests running on the new objects.